fastapi/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
boto3==1.33.13
botocore==1.33.13
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
loguru==0.7.2
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-magic==0.4.27
s3transfer==0.8.2
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.27.0.post1
python-multipart==0.0.6
python-dotenv==1.0.1